requirements.txt
customtkinter
pyglet
matplotlib
pywinstyles
python-dateutil